Feature #5157 : - upgrading galleria slider plugin - wiring the player (video / audio) Silverpeas JQuery plugin - adding sound view and edit pages - adding sound labels - modifying message on empty albums - modifying the model to manipulate a definition object instead of width and height integers - improving the browse-bar content when editing a media - handling the save of definition and duration of a video - handling the save of duration of a sound - adding modify menu item into media view pages - adding sound processor method into media helper - renaming bundle key (for centralized treatments) - renaming some classes to use media word rather than image - moving metadata management from Photo model to Internal media model (even if only photo use it for now) - changing the way to get file resources into some of unit tests in order to execute them successfully from IDEs (without using maven)

+++ gallery/gallery-ejb/src/main/java/com/silverpeas/gallery/process/media/GalleryUpdateMediaFileProcess.java
@@ -87,38 +87,35 @@ public class GalleryUpdateMediaFileProcess extends AbstractGalleryFileProcess {
   public void processFiles(final GalleryProcessExecutionContext context,
       final ProcessSession session, final FileHandler fileHandler) throws Exception {
     // Media
-    switch (getMedia().getType()) {
-      case Photo:
-        if (fileItem != null) {
-          final String name = fileItem.getName();
-          if (StringUtil.isDefined(name)) {
+    if (fileItem != null && !getMedia().getType().isStreaming()) {
+      final String name = fileItem.getName();
+      if (StringUtil.isDefined(name)) {
 
-            // Deleting repository with old media
-            fileHandler.getHandledFile(Media.BASE_PATH, context.getComponentInstanceId(),
-                getMedia().getWorkspaceSubFolderName()).delete();
+        // Deleting repository with old media
+        fileHandler.getHandledFile(Media.BASE_PATH, context.getComponentInstanceId(),
+            getMedia().getWorkspaceSubFolderName()).delete();
 
+        switch (getMedia().getType()) {
+          case Photo:
             // Creating new images
             MediaHelper
                 .processPhoto(fileHandler, getMedia().getPhoto(), fileItem, watermark, watermarkHD,
                     watermarkOther);
-          }
+            break;
+          case Video:
+            // Save new video
+            MediaHelper.processVideo(fileHandler, getMedia().getVideo(), fileItem);
+            break;
+          case Sound:
+            // Save new sound
+            MediaHelper.processSound(fileHandler, getMedia().getSound(), fileItem);
+            break;
+          default:
+            SilverTrace.warn("Gallery", GalleryUpdateMediaFileProcess.class.getName(),
+                getMedia().getType().name() + " media type is never processed");
+            break;
         }
-        break;
-      case Video:
-        if (fileItem != null) {
-          // Deleting repository with old media
-          fileHandler.getHandledFile(Media.BASE_PATH, context.getComponentInstanceId(),
-              getMedia().getWorkspaceSubFolderName()).delete();
-
-          // Save new video
-          MediaHelper.processVideo(fileHandler, getMedia().getVideo(), fileItem);
-        }
-        break;
-      default:
-        SilverTrace.warn("Gallery", GalleryUpdateMediaFileProcess.class.getName(),
-            getMedia().getType().name() + " media type is never processed");
-        break;
+      }
     }
-
   }
 }
